refactor(home): extracted some components from home
